compiler/clib: ULONG->unsigned int for sigaddset(), sigdelset() and sigismember()
git-svn-id: https://svn.aros.org/svn/aros/trunk@44122 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/sigismember.c
@@ -5,12 +5,11 @@
     ANSI C function sigismember().
 */
 
+#include <stddef.h>
+
 /*****************************************************************************
 
     NAME */
-
-#include "__arosc_privdata.h"
-
 #include <signal.h>
 
 	int sigismember (
@@ -38,11 +37,11 @@
 ******************************************************************************/
 {
     if (NULL != set) {
-        ULONG i = (signum >> 5);
+        unsigned int i = (signum >> 5);
         if (0 != (set->__val[i] & (signum & 0x1f)))
             return 1;
         return 0;
     }
-	
+
     return -1;
 } /* sigismember */
